+++ CHANGELOG.md
@@ -11,6 +11,7 @@
   - The int `0` was not a correct value for the function getPrimaryKeyValue. As a restult, store collections could have duplicate values.
   - The validation did not return the callback for a the store with no map.
   - The directive `v-lunaris` did not work with radio values.
+  - The store collection did not delete index values.
 
 ## 1.6.1
 *2018-11-19*